handle slack error 429
This modifies the behavior when an error `429 - Too Many Requests` occur
when it sends a request to slack. It now prints a log instead of raising
an error.

Before that, the log that occured the failure was pushed in a "pending"
queue. This queue doesn't wait long enough to send another request to
slack, resulting in another `429`.